# pixela-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixela API Documentation",
        "description": "API Documentation for Pixela - A platform for tracking and managing your movie and series watching history",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Listar reseñas del usuario",
                "description": "Obtiene todas las reseñas del usuario autenticado",
                "operationId": "listReviews",
                "responses": {
                    "200": {
                        "description": "Lista de reseñas recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Añadir una nueva reseña",
                "description": "Crea una nueva reseña para una película o serie",
                "operationId": "addReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reseña creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "La reseña ya existe o datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reviews/{review}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Actualizar una reseña",
                "description": "Actualiza una reseña existente",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "ID de la reseña",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reseña actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para actualizar esta reseña",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Eliminar una reseña",
                "description": "Elimina una reseña específica",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "ID de la reseña",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reseña eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Review deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para eliminar esta reseña",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "General error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    }
                },
                "type": "object"
            },
...
            "ReviewListResponse": {
                "description": "List of reviews response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Reviews retrieved successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    }
                },
                "type": "object"
            },
            "ReviewRequest": {
                "description": "Data required to create a review",
                "required": [
                    "item_type",
                    "tmdb_id",
                    "rating"
                ],
                "properties": {
                    "item_type": {
                        "type": "string",
                        "enum": [
                            "movie",
                            "series"
                        ],
                        "example": "movie"
                    },
                    "tmdb_id": {
                        "type": "integer",
                        "example": 550
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "maximum": 10,
                        "minimum": 0,
                        "example": 8.5
                    },
                    "review": {
                        "type": "string",
                        "example": "An excellent movie...",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ReviewResponse": {
                "description": "Single review response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Review operation successful"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Review"
                    }
                },
                "type": "object"
            },
            "Review": {
                "description": "Review information",
                "required": [
                    "id",
                    "user_id",
                    "tmdb_id",
                    "item_type",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "photo_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "tmdb_id": {
                        "type": "integer",
                        "example": 550
                    },
                    "item_type": {
                        "type": "string",
                        "enum": [
                            "movie",
                            "series"
                        ],
                        "example": "movie"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "maximum": 10,
                        "minimum": 0,
                        "example": 8.5
                    },
                    "review": {
                        "type": "string",
                        "example": "An excellent movie...",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "title": {
                        "type": "string",
                        "example": "Fight Club"
                    },
                    "poster_path": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ReviewUpdateRequest": {
                "description": "Data required to update a review",
                "required": [
                    "rating"
                ],
                "properties": {
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "maximum": 10,
                        "minimum": 0,
                        "example": 8.5
                    },
                    "review": {
                        "type": "string",
                        "example": "An excellent movie...",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum. Include the token in the Authorization header as 'Bearer {token}'",
                "scheme": "bearer"
            }
        }
    },
...
}